requirements.txt
anthropic>=0.25.0
pypdf>=4.0.0
pdfplumber>=0.10.0
reportlab>=4.0.0
python-docx>=1.1.0
requests>=2.31.0
python-dotenv>=1.0.0
notion-client>=2.2.0
rich>=13.7.0
typer>=0.12.0
httpx>=0.27.0
flask>=3.0.0
google-generativeai>=0.8.0
